Multiple minor updates

+++ src/main/java/fr/theoszanto/mc/express/utils/Weighted.java
@@ -1,5 +1,5 @@
 package fr.theoszanto.mc.express.utils;
 
 public interface Weighted {
-	int getWeight();
+	double getWeight();
 }
